chore: renamed options.ts to context
