+++ src/http/path.ts
@@ -1,4 +1,4 @@
-import { Context } from '../options';
+import { Context } from '../context';
 
 export default class Path {
   private readonly path: string;
